backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.33.0
httpx>=0.27.0,<0.29.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0

langchain-core>=0.3.28,<0.4.0
langchain-google-genai>=2.0.7,<3.0.0
langchain-experimental>=0.3.4,<0.4.0

numpy>=2.0.0,<3.0.0
matplotlib>=3.9.0,<4.0.0
